LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-mutation-replacesync-expected.txt


FAIL Fetched sheet replaces placeholder including replaceSync content. assert_equals: Before replaceSync: expected 1 placeholder(s). expected 1 but got 0
